Test script for paylater button eligibility validation for any geolocation along with script usage instruction md file
